♻️ refactor: problemInfo 타입이 있는 모듈을 옮기고, 영향을 받는 모듈들의 경로를 업데이트
