+++ types/gacha.ts
@@ -1,7 +1,2 @@
 import type { Tier } from './randomDefense';
 
-export interface ProblemInfo {
-  problemId: number;
-  title: string;
-  tier: Tier;
-}
